complete refonte architecture
