+++ config/tests.py
@@ -1,0 +1,2 @@
+def test_dummy():
+    assert 1
