--- oc_lettings_site/asgi.py
@@ -1,7 +1,0 @@
-import os
-
-from django.core.asgi import get_asgi_application
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'oc_lettings_site.settings')
-
-application = get_asgi_application()
