+++ profiles/admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+# Register your models here.
+from .models import Profile
+
+admin.site.register(Profile)
+
+
