+++ setup.cfg
@@ -3,6 +3,6 @@ max-line-length = 99
 exclude = **/migrations/*,venv
 
 [tool:pytest]
-DJANGO_SETTINGS_MODULE = oc_lettings_site.settings
+DJANGO_SETTINGS_MODULE = config.settings
 python_files = tests.py
 addopts = -v
